[Fix] UserBlockRequestDTO에 blockedUserId 타입 변경

+++ travelPlan/Source/Data/Network/RequestDTO/UserBlockRequestDTO.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 struct UserBlockRequestDTO: Encodable {
-  let blockedUserId: Int32
+  let blockedUserId: Int64
 }
